+++ gazebo/physics/World.cc
@@ -750,6 +750,13 @@ void World::Update()
     }
   }
 
+  // Give clients a possibility to react to collisions before the physics
+  // gets updated.
+  this->dataPtr->updateInfo.realTime = this->GetRealTime();
+  event::Events::beforePhysicsUpdate(this->dataPtr->updateInfo);
+
+  DIAG_TIMER_LAP("World::Update", "Events::beforePhysicsUpdate");
+
   // Update the physics engine
   if (this->dataPtr->enablePhysicsEngine && this->dataPtr->physicsEngine)
   {
@@ -2076,31 +2083,6 @@ void World::SetState(const WorldState &_state)
   this->dataPtr->logRealTime = _state.GetRealTime();
   this->dataPtr->iterations = _state.GetIterations();
 
-  // Models
-  const ModelState_M modelStates = _state.GetModelStates();
-  for (auto const &modelState : modelStates)
-  {
-    ModelPtr model = this->GetModel(modelState.second.GetName());
-    if (model)
-      model->SetState(modelState.second);
-    else
-      gzerr << "Unable to find model[" << modelState.second.GetName() << "]\n";
-  }
-
-  // Lights
-  const LightState_M lightStates = _state.LightStates();
-  for (auto const &lightState : lightStates)
-  {
-    LightPtr light = this->Light(lightState.second.GetName());
-    if (light)
-      light->SetState(lightState.second);
-    else
-    {
-      gzerr << "Unable to find light[" << lightState.second.GetName() << "]"
-            << std::endl;
-    }
-  }
-
   // Insertions (adapted from ProcessFactoryMsgs)
   auto insertions = _state.Insertions();
   for (auto const &insertion : insertions)
@@ -2180,6 +2162,39 @@ void World::SetState(const WorldState &_state)
             std::endl;
       }
     }
+  }
+
+  // Model updates
+  const ModelState_M modelStates = _state.GetModelStates();
+  for (auto const &modelState : modelStates)
+  {
+    ModelPtr model = this->GetModel(modelState.second.GetName());
+    if (model)
+      model->SetState(modelState.second);
+    else
+      gzerr << "Unable to find model[" << modelState.second.GetName() << "]\n";
+  }
+
+  // Light updates
+  const LightState_M lightStates = _state.LightStates();
+  for (auto const &lightState : lightStates)
+  {
+    LightPtr light = this->Light(lightState.second.GetName());
+    if (light)
+      light->SetState(lightState.second);
+    else
+    {
+      gzerr << "Unable to find light[" << lightState.second.GetName() << "]"
+            << std::endl;
+    }
+  }
+
+  // Deletions
+  auto deletions = _state.Deletions();
+  for (auto const &deletion : deletions)
+  {
+    // This works for models and lights
+    this->RemoveModel(deletion);
   }
 }